renaming `user.py` to `users.py`
